fix: run_flow_from_json circular dependency
